fix: wrong recomposition when open dialog or show scroll to top button
