+++ app/src/main/java/io/github/skydynamic/maiproberplus/ui/compose/scores/maimai/MaimaiScoreList.kt
@@ -1,6 +1,10 @@
 package io.github.skydynamic.maiproberplus.ui.compose.scores.maimai
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,15 +13,16 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
-import androidx.compose.foundation.lazy.grid.LazyGridState
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedTextField
@@ -30,16 +35,23 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import io.github.skydynamic.maiproberplus.R
 import io.github.skydynamic.maiproberplus.core.data.maimai.MaimaiScoreManager.deleteAllScore
 import io.github.skydynamic.maiproberplus.ui.compose.ConfirmDialog
 import io.github.skydynamic.maiproberplus.ui.compose.scores.ScoreManagerViewModel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 
 @Composable
-fun MaimaiScoreList(gridState: LazyGridState) {
+fun MaimaiScoreList(
+    coroutineScope: CoroutineScope
+) {
     var openDeleteConfirmDialog by remember { mutableStateOf(false) }
     var openCreateScoreDialog by remember { mutableStateOf(false) }
+    val gridState = rememberLazyGridState()
 
     when {
         openDeleteConfirmDialog -> {
@@ -61,6 +73,16 @@ fun MaimaiScoreList(gridState: LazyGridState) {
                     openCreateScoreDialog = false
                 }
             )
+        }
+        ScoreManagerViewModel.showMaimaiScoreSelectionDialog -> {
+            if (ScoreManagerViewModel.maimaiScoreSelection == null) {
+                ScoreManagerViewModel.showMaimaiScoreSelectionDialog = false
+            } else {
+                MaimaiScoreDetailDialog(ScoreManagerViewModel.maimaiScoreSelection!!) {
+                    ScoreManagerViewModel.showMaimaiScoreSelectionDialog = false
+                    ScoreManagerViewModel.maimaiScoreSelection = null
+                }
+            }
         }
     }
 
@@ -135,13 +157,13 @@ fun MaimaiScoreList(gridState: LazyGridState) {
         ) {
             Row(
                 modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(8.dp),
+                    .padding(4.dp)
+                    .fillMaxWidth(),
                 horizontalArrangement = Arrangement.spacedBy(8.dp),
                 verticalAlignment = Alignment.CenterVertically,
             ) {
                 Button(
-                    modifier = Modifier.padding(4.dp).weight(0.45f),
+                    modifier = Modifier.weight(1f),
                     onClick = {
                         openCreateScoreDialog = true
                     }
@@ -151,7 +173,7 @@ fun MaimaiScoreList(gridState: LazyGridState) {
                 }
 
                 Button(
-                    modifier = Modifier.padding(4.dp).weight(0.55f),
+                    modifier = Modifier.weight(1f),
                     onClick = {
                         openDeleteConfirmDialog = true
                     },
@@ -180,6 +202,33 @@ fun MaimaiScoreList(gridState: LazyGridState) {
                     ScoreManagerViewModel.showMaimaiScoreSelectionDialog = true
                 }
             )
+        }
+    }
+
+    AnimatedVisibility(
+        visible = gridState.canScrollBackward,
+        enter = fadeIn(),
+        exit = fadeOut()
+    ) {
+        Box(
+            Modifier
+                .fillMaxSize()
+        ) {
+            FloatingActionButton(
+                onClick = {
+                    coroutineScope.launch {
+                        gridState.scrollToItem(0)
+                    }
+                },
+                modifier = Modifier
+                    .align(Alignment.BottomEnd)
+                    .padding(16.dp)
+            ) {
+                Icon(
+                    painterResource(R.drawable.arrow_upward_24px),
+                    contentDescription = null
+                )
+            }
         }
     }
 }
